requirements.txt
seaborn
scipy
pandas
matplotlib
torch
torchvision
hal-cgp
opencv-python